Refactored user objects and updated user builders.
