+++ src/main/java/com/smartsheet/api/models/User.java
@@ -28,6 +28,11 @@ package com.smartsheet.api.models;
  */
 public class User extends UserProfile {
 	/**
+	 * Represents the name
+	 */
+	private String name;
+
+	/**
 	 * Represents the admin flag which allows managing users and accounts.
 	 */
 	private Boolean admin;
@@ -37,22 +42,37 @@ public class User extends UserProfile {
 	 */
 	private Boolean licensedSheetCreator;
 
-	/**
-	 * Represents the resource manager flag which allows the user access to the Resource Manager functionality.
-	 */
-	private Boolean resourceViewer; // TODO: resourceViewer or resourceManager?
+	// TODO: Determine if this value should be resourceViewer or resourceManager. Comment out until then.
+//	/**
+//	 * Represents the resource manager flag which allows the user access to the Resource Manager functionality.
+//	 */
+//	private Boolean resourceViewer;
 	
 	/**
 	 * Represents the group admin flag which allows users to create and modify groups.
 	 */
 	private Boolean groupAdmin;
-
-	
 	
 	/**
 	 * Represents the user status (active, pending, declined).
 	 */
 	private UserStatus status;
+
+	/**
+	 * Gets the name
+	 * @return the name
+	 */
+	public String getName() {
+		return name;
+	}
+
+	/**
+	 * Sets the name
+	 * @param name the name
+	 */
+	public void setName(String name) {
+		this.name = name;
+	}
 
 	/**
 	 * Gets the admin flag which allows managing users and accounts.
@@ -108,20 +128,21 @@ public class User extends UserProfile {
 		this.status = status;
 	}
 
-	/**
-	 * @return the flag indicating if someone is a resource manager
-	 */
-	public Boolean getResourceViewer() {
-		return resourceViewer;
-	}
-
-	/**
-	 * Sets the resource manager flag.
-	 * @param resourceViewer
-	 */
-	public void setResourceViewer(Boolean resourceViewer) {
-		this.resourceViewer = resourceViewer;
-	}
+	// TODO: Determine if this value should be resourceViewer or resourceManager. Comment out until then.
+//	/**
+//	 * @return the flag indicating if someone is a resource manager
+//	 */
+//	public Boolean getResourceViewer() {
+//		return resourceViewer;
+//	}
+//
+//	/**
+//	 * Sets the resource manager flag.
+//	 * @param resourceViewer
+//	 */
+//	public void setResourceViewer(Boolean resourceViewer) {
+//		this.resourceViewer = resourceViewer;
+//	}
 
 	/**
 	 * 
@@ -145,7 +166,7 @@ public class User extends UserProfile {
 		private String emailAddress;
 		private Boolean licensedSheetCreator;
 		private Boolean groupAdmin;
-		private Boolean resourceViewer;
+		//private Boolean resourceViewer; // TODO: Determine if this value should be resourceViewer or resourceManager. Comment out until then.
 		private String firstName;
 		private String lastName;
 
@@ -245,20 +266,21 @@ public class User extends UserProfile {
 			return this;
 		}
 
-		/**
-		 * @return the resourceViewer
-		 */
-		public Boolean getResourceViewer() {
-			return resourceViewer;
-		}
-
-		/**
-		 * @param resourceViewer the resourceViewer to set
-		 */
-		public AddUserBuilder setResourceViewer(Boolean resourceViewer) {
-			this.resourceViewer = resourceViewer;
-			return this;
-		}
+		// TODO: Determine if this value should be resourceViewer or resourceManager. Comment out until then.
+//		/**
+//		 * @return the resourceViewer
+//		 */
+//		public Boolean getResourceViewer() {
+//			return resourceViewer;
+//		}
+//
+//		/**
+//		 * @param resourceViewer the resourceViewer to set
+//		 */
+//		public AddUserBuilder setResourceViewer(Boolean resourceViewer) {
+//			this.resourceViewer = resourceViewer;
+//			return this;
+//		}
 		/**
 		 * Builds the {@link User} object using the required fields.
 		 *
@@ -273,7 +295,7 @@ public class User extends UserProfile {
 			user.admin = admin;
 			user.licensedSheetCreator = licensedSheetCreator;
 			user.groupAdmin = groupAdmin;
-			user.resourceViewer = resourceViewer;
+			//user.resourceViewer = resourceViewer; // TODO: Determine if this value should be resourceViewer or resourceManager. Comment out until then.
 			user.setFirstName(firstName);
 			user.setLastName(lastName);
 			user.setEmail(emailAddress);
@@ -290,27 +312,9 @@ public class User extends UserProfile {
 		private Boolean admin;
 		private Boolean licensedSheetCreator;
 		private Boolean groupAdmin;
-		private Boolean resourceViewer;
-		private Long id;
+		//private Boolean resourceViewer; // TODO: Determine if this value should be resourceViewer or resourceManager. Comment out until then.
 		private String firstName;
 		private String lastName;
-
-		/**
-		 * Get the id of the user
-		 * @return
-		 */
-		public Long getId() {
-			return id;
-		}
-
-		/**
-		 * Set the id of the user
-		 * @param id
-		 */
-		public UpdateUserBuilder setId(Long id) {
-			this.id = id;
-			return this;
-		}
 
 		/**
 		 * Sets the admin flag which allows managing users and accounts.
@@ -397,20 +401,21 @@ public class User extends UserProfile {
 			return this;
 		}
 
-		/**
-		 * @return the resourceViewer
-		 */
-		public Boolean getResourceViewer() {
-			return resourceViewer;
-		}
-
-		/**
-		 * @param resourceViewer the resourceViewer to set
-		 */
-		public UpdateUserBuilder setResourceViewer(Boolean resourceViewer) {
-			this.resourceViewer = resourceViewer;
-			return this;
-		}
+		// TODO: Determine if this value should be resourceViewer or resourceManager. Comment out until then.
+//		/**
+//		 * @return the resourceViewer
+//		 */
+//		public Boolean getResourceViewer() {
+//			return resourceViewer;
+//		}
+//
+//		/**
+//		 * @param resourceViewer the resourceViewer to set
+//		 */
+//		public UpdateUserBuilder setResourceViewer(Boolean resourceViewer) {
+//			this.resourceViewer = resourceViewer;
+//			return this;
+//		}
 		
 		/**
 		 * Builds the {@link User}.
@@ -423,13 +428,12 @@ public class User extends UserProfile {
 			}
 			
 			User user = new User();
-			user.setId(id);
 			user.setFirstName(firstName);
 			user.setLastName(lastName);
 			user.admin = admin;
 			user.licensedSheetCreator = licensedSheetCreator;
 			user.groupAdmin = groupAdmin;
-			user.resourceViewer = resourceViewer;
+			//user.resourceViewer = resourceViewer; // TODO: Determine if this value should be resourceViewer or resourceManager. Comment out until then.
 			return user;
 		}
